evalelevance/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
httpx
python-dotenv 